Fix the size of a strncpy
It seems that the intent of this code was to only use the first char[24]
in `addrs` for this strncpy. While that should always happen anyway
(inet_ntoa should never produce a >= 24 char string), the mismatch
between the size of poop->addrs[0] and `sizeof(poop->addrs)` caused
issues with FORTIFY on clang.